Reclassify units and re-structure source files.
Units are now classified into Reach-Forming, Structure, Boundary or Junction.
